알터창 바꿔서 커밋

+++ src/main/java/com/project/controller/WishController.java
@@ -69,16 +69,28 @@ public class WishController {
 	}
 	@RequestMapping(value="/deleteAllWish.action",method= {RequestMethod.POST})
 	@ResponseBody
-	public String deleteAllWish(HttpServletRequest req)throws Exception{
-		
+	public HashMap<String, Object> deleteAllWish(HttpServletRequest req)throws Exception{
+		System.out.println("------------------111111111111111");
+		HashMap<String, Object> map =new HashMap<String, Object>();
 		HttpSession info = req.getSession();
 		MemberDTO m_dto = (MemberDTO) info.getAttribute("userInfo");
+		System.out.println("------------------222222222222222222");
 		String m_id = m_dto.getM_ID();
+		System.out.println(m_id);
+		int idDataCount=w_service.idDataCount(m_id);
+		System.out.println("------------------3333333333333333");
+		System.out.println(idDataCount);
+		if(idDataCount !=0) {
 		w_service.deleteAllWish(m_id);
+		map.put("msg","찜목록이 삭제되었습니다");
+		
+		}else {
+		map.put("msg","찜목록이 없습니다");	
+			
+		}
 		
 		
-		
-		return "success";
+		return map;
 		
 	}
 	@RequestMapping(value="/wishCheck.action",method= {RequestMethod.POST})
@@ -90,8 +102,6 @@ public class WishController {
 		String m_id = m_dto.getM_ID();
 		HashMap<String, Object> map = new HashMap<String, Object>();
 		int count = w_service.checkWish(G_NUM, m_id);//데이터가 있는지 없는지 확인
-		System.out.println("----------카운트카운트~~~--------------------");
-		System.out.println(count);
 		map.put("count",count);
 		map.put("G_NUM",G_NUM);
